Incorporate suggestions
