+++ brainsets/datasets/AllenVisualCodingOphys2016.py
@@ -3,7 +3,7 @@ from pathlib import Path
 
 from torch_brain.dataset import Dataset, CalciumImagingDatasetMixin
 
-from brainsets.config import get_processed_dir
+from ._utils import get_processed_dir
 
 
 class AllenVisualCodingOphys2016(CalciumImagingDatasetMixin, Dataset):
